admin/etl/requirements.txt
flake8==6.1.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
psycopg2==2.9.9
elasticsearch==8.11.1
